After new update and forum working

+++ application/helpers/multi_language_helper.php
@@ -22,15 +22,39 @@ if ( ! function_exists('get_phrase'))
         $CI->load->database();
         $language_code = $CI->db->get_where('settings' , array('key' => 'language'))->row()->value;
         $key = strtolower(preg_replace('/\s+/', '_', $phrase));
-
         $langArray = openJSONFile($language_code);
-        if (array_key_exists($key, $langArray)) {
-        } else {
+
+        // THIS BLOCK OF CODE IS THE CORE FOR TRANSLATING
+        if ($langArray && !array_key_exists($key, $langArray) ) {
+            $langArray[$key] = ucfirst(str_replace('_', ' ', $key));
+            $jsonData = json_encode($langArray, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE);
+
+            file_put_contents(APPPATH.'language/'.$language_code.'.json', stripslashes($jsonData));
+        }
+        // THIS BLOCK OF CODE IS THE CORE FOR TRANSLATING
+        return $langArray[$key];
+    }
+}
+
+// This function helps us to get the translated phrase from the file. If it does not exist this function will save the phrase and by default it will have the same form as given
+if ( ! function_exists('site_phrase'))
+{
+    function site_phrase($phrase = '') {
+        $CI	=&	get_instance();
+        if (!$CI->session->userdata('language')) {
+            $CI->session->set_userdata('language', get_settings('language'));
+        }
+        $language_code = $CI->session->userdata('language');
+        $key = strtolower(preg_replace('/\s+/', '_', $phrase));
+        $langArray = openJSONFile($language_code);
+
+        // THIS BLOCK OF CODE IS THE CORE FOR TRANSLATING
+        if ($langArray && !array_key_exists($key, $langArray) ) {
             $langArray[$key] = ucfirst(str_replace('_', ' ', $key));
             $jsonData = json_encode($langArray, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE);
             file_put_contents(APPPATH.'language/'.$language_code.'.json', stripslashes($jsonData));
         }
-
+        // THIS BLOCK OF CODE IS THE CORE FOR TRANSLATING
         return $langArray[$key];
     }
 }
@@ -54,11 +78,7 @@ if ( ! function_exists('saveDefaultJSONFile'))
 {
     function saveDefaultJSONFile($language_code){
         $language_code = strtolower($language_code);
-        if(file_exists(APPPATH.'language/'.$language_code.'.json')){
-            $newLangFile 	= APPPATH.'language/'.$language_code.'.json';
-            $enLangFile   = APPPATH.'language/english.json';
-            copy($enLangFile, $newLangFile);
-        }else {
+        if(!file_exists(APPPATH.'language/'.$language_code.'.json')){
             $fp = fopen(APPPATH.'language/'.$language_code.'.json', 'w');
             $newLangFile = APPPATH.'language/'.$language_code.'.json';
             $enLangFile   = APPPATH.'language/english.json';
@@ -76,14 +96,29 @@ if ( ! function_exists('saveJSONFile'))
         if(file_exists(APPPATH.'language/'.$language_code.'.json')){
             $jsonString = file_get_contents(APPPATH.'language/'.$language_code.'.json');
             $jsonString = json_decode($jsonString, true);
-            $jsonString[$updating_key] = $updating_value;
+            $jsonString[$updating_key] = filter_var(escapeJsonString($updating_value), FILTER_SANITIZE_SPECIAL_CHARS, FILTER_SANITIZE_STRING);
         }else {
-            $jsonString[$updating_key] = $updating_value;
+            $jsonString[$updating_key] = filter_var(escapeJsonString($updating_value), FILTER_SANITIZE_SPECIAL_CHARS, FILTER_SANITIZE_STRING);
         }
         $jsonData = json_encode($jsonString, JSON_PRETTY_PRINT|JSON_UNESCAPED_UNICODE);
         file_put_contents(APPPATH.'language/'.$language_code.'.json', stripslashes($jsonData));
     }
 }
+
+
+// This function helps us to update a phrase inside the language file.
+if ( ! function_exists('escapeJsonString'))
+{
+    function escapeJsonString($value) {
+        $value = str_replace('"', "'", $value);
+        $escapers =     array("\\",     "/",   "\"",  "\n",  "\r",  "\t", "\x08", "\x0c");
+        $replacements = array("\\\\", "\\/", "\\\"", "\\n", "\\r", "\\t",  "\\f",  "\\b");
+        $result = str_replace($escapers, $replacements, $value);
+        return $result;
+    }
+}
+
+
 
 
 // ------------------------------------------------------------------------
